rauc-hawkbit: update license hash for README
The README has changed to indicated that rauc-hawkbit is not meant for
production use, moving lines around.

Interestingly the line numbers in the recipe actually seem to correct
for the new README, but the checksum is not.

Update the checksum to match. The relevant part of the README did not
change.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>